requirements.txt
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cryptography==44.0.2
distlib==0.3.9
filelock==3.17.0
identify==2.6.8
idna==3.10
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.1.0
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
virtualenv==20.29.2
